+++ src/maicos/modules/densityplanar.py
@@ -72,7 +72,7 @@ class DensityPlanar(ProfilePlanarBase):
             bin_width=bin_width,
             refgroup=refgroup,
             sym=sym,
-            sym_parity="even",
+            sym_odd=False,
             grouping=grouping,
             bin_method=bin_method,
             output=output,